full-text search, improved admin pages, ban system, warn system, maintenance mode, forum permission system, code cleanup, added missing LICENSE/COPYING files

+++ forums/views.py
@@ -1,34 +1,40 @@
 import math
 from datetime import datetime, timedelta
-from django.shortcuts import redirect, get_object_or_404, get_list_or_404
+from django.shortcuts import (redirect, render_to_response, get_object_or_404)
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.core.cache import cache
 from django.core.urlresolvers import reverse
-from django.db.models import F, Q
+from django.db.models import F, Q, Count
 from django.template import RequestContext
 from forums.forms import (PostForm, DeletePostForm, TopicForm,
-	MoveTopicForm, PostSearchForm, ReportPostForm)
+	MoveTopicForm, ReportPostForm)
 from forums.models import Category, Forum, Topic, Post, PostKarma, Report
 from utilities.annoying.decorators import render_to
 from utilities.annoying.functions import get_config, get_object_or_None
-from utilities.internal.decorators import has_perm_or_403, user_passes_test_or_403
-from utilities.internal.templatetags.forums import editable_by
+from utilities.internal.decorators import user_passes_test_or_403
+from utilities.internal.templatetags.forums import (can_access_forum,
+	can_post_topic, can_post_reply)
 
 
 @render_to('forums/index.html')
 def index(request):
 	users_online = cache.get('forums_users_online')
-	if users_online == None:		
+	if users_online == None:
 		users_online = list(User.objects.filter(profile__last_activity_at__gte=
 			(datetime.now()-timedelta(minutes=15))))
 		cache.set('forums_users_online', users_online)
-	
-	categories = cache.get('forums_categories')
+
+	categories = cache.get('forums_categories_%s' % request.user.id
+		if request.user.is_authenticated() else 'anon')
 	if categories == None:
-		forums = list(Forum.objects.all(). \
+		#forums = list(Forum.objects.all(). \
+		#	select_related('category', 'last_post__topic', 'last_post__author'))
+		forums = list(Forum.objects.annotate(num_can_read=Count('can_read')). \
 			select_related('category', 'last_post__topic', 'last_post__author'))
+		forums = [i for i in forums
+			if can_access_forum(request, i, return_plain_boolean=True)]
 		categories = {}
 		for forum in forums:
 			cat = categories.setdefault(forum.category.id,
@@ -36,7 +42,8 @@ def index(request):
 			cat['forums'].append(forum)
 		cmpdef = lambda a, b: cmp(a['category'].order, b['category'].order)
 		categories = sorted(categories.values(), cmpdef)
-		cache.set('forums_categories', categories)
+		cache.set('forums_categories_%s' % request.user.id
+			if request.user.is_authenticated() else 'anon', categories)
 
 	posts = cache.get('forums_count_posts')
 	if posts == None:
@@ -53,7 +60,6 @@ def index(request):
 		users = User.objects.count()
 		cache.set('forums_count_users', users)
 
-
 	return {'categories': categories, 'users_online': users_online,
 			'posts': posts, 'topics': topics, 'users': users}
 
@@ -62,14 +68,18 @@ def index(request):
 def forum_view(request, forum_id):
 	forum = cache.get('forums_forum_%s' % forum_id)
 	if forum == None:
-		forum = get_object_or_404(Forum, pk=forum_id)
+		forum = get_object_or_404(Forum.objects \
+			.annotate(num_can_read=Count('can_read')), pk=forum_id)
 		cache.set('forums_forum_%s' % forum_id, forum)
-	topics = cache.get('forum_topics_%s' % forum_id)
+	can_access = can_access_forum(request, forum)
+	if not can_access == True:
+		return can_access
+	topics = cache.get('forums_topics_%s' % forum_id)
 	if topics == None:
 		topics = list(Topic.objects.filter(forum__pk=forum_id). \
 			order_by('-is_sticky', '-created_at'). \
 			select_related('author', 'last_post__author'))
-		cache.set('forum_topics_%s' % forum_id, topics)
+		cache.set('forums_topics_%s' % forum_id, topics)
 	return {'forum': forum, 'topics': topics}
 
 
@@ -79,12 +89,16 @@ def topic_view(request, topic_id):
 	if topic == None:
 		topic = get_object_or_404(Topic.objects.select_related(), pk=topic_id)
 		cache.set('forums_topic_%s' % topic_id, topic)
+	can_access = can_access_forum(request, topic.forum)
+	if not can_access == True:
+		return can_access
 	if request.user.is_authenticated():
 		topic.update_read(request.user)
 	Topic.objects.filter(pk=topic_id).update(view_count=F('view_count') + 1) # FIXME ++ with cache
 	posts = cache.get('forums_posts_%s' % topic_id)
 	if posts == None:
-		posts = list(Post.objects.filter(topic__pk=topic_id).select_related('author__profile__badge'))
+		posts = list(Post.objects.filter(topic__pk=topic_id) \
+			.select_related('author__profile__group', 'karma'))
 		cache.set('forums_posts_%s' % topic_id, posts)
 	return {'topic': topic, 'posts': posts, 'first_post_id': posts[0].id}
 
@@ -95,19 +109,17 @@ def post_permalink(request, post_id):
 		created_at__lt=post.created_at).count()
 	page = int(math.ceil((float(older_posts) + 1.0) /
 		get_config('POSTS_PER_PAGE', 25)))
-	return redirect(reverse('forums.views.topic_view',
-		kwargs={'topic_id': post.topic.id}) + '?page=%s#post-%s' % (page, post.id))
+	return redirect(post.topic.get_abolute_url() + '?page=%s#post-%s' % (page, post.id))
 
 
-@has_perm_or_403('forums.add_topic')
+@login_required
 @render_to('forums/topic_new.html')
 def topic_new(request, forum_id):
 	forum = get_object_or_404(Forum, pk=forum_id)
-	if not forum.can_regular_user_post and not request.user.is_staff:
-		messages.error(request, "You are not allowed to post new topics \
+	if not can_post_topic(request.user, forum):
+		messages.error(request, "You are not allowed to to post new topics \
 			on this forum.")
-		return redirect(reverse('forums.views.forum_view', 
-			kwargs={'forum_id': forum.id}))
+		return redirect(forum.get_abolute_url())
 	if request.method == 'POST':
 		form = TopicForm(request.POST)
 		if form.is_valid():
@@ -115,7 +127,8 @@ def topic_new(request, forum_id):
 			content = form.cleaned_data['content']
 			topic = Topic(title=title, author=request.user, forum=forum)
 			topic.save()
-			post = Post(topic=topic, author=request.user, content=content)
+			post = Post(topic=topic, author=request.user,
+				author_ip=request.META['REMOTE_ADDR'], content=content)
 			post.save()
 			messages.success(request, "Your topic has been saved.")
 			return redirect(topic.get_absolute_url())
@@ -124,18 +137,22 @@ def topic_new(request, forum_id):
 	return {'forum': forum, 'form': form}
 
 
-@has_perm_or_403('forums.add_post')
+@login_required
 @render_to('forums/post_new.html')
 def post_new(request, topic_id, quoted_post_id=None):
-	topic = get_object_or_404(Topic, pk=topic_id)
+	topic = get_object_or_404(Topic.objects.select_related(), pk=topic_id)
 	if topic.is_closed and not request.user.is_staff:
 		messages.error(request, "You are not allowed to post in closed topics.")
+		return redirect(topic.get_absolute_url())
+	if not can_post_reply(request.user, topic.forum):
+		messages.error(request, "You are not allowed to reply on this forum.")
 		return redirect(topic.get_absolute_url())
 	if request.method == 'POST':
 		form = PostForm(request.POST)
 		if form.is_valid():
 			content = form.cleaned_data['content']
-			post = Post(topic=topic, author=request.user, content=content)
+			post = Post(topic=topic, author=request.user,
+				author_ip=request.META['REMOTE_ADDR'], content=content)
 			post.save()
 			messages.success(request, "Your reply has been saved.")
 			return redirect(post.get_absolute_url())
@@ -147,22 +164,24 @@ def post_new(request, topic_id, quoted_post_id=None):
 				form.initial = {'content': "[quote=%s]%s[/quote]" %
 					(quoted_post.author, quoted_post.content)}
 			except Post.DoesNotExist:
-				messages.warning(request, "You tried to quote a post which doesn't \
-					exist or it doesn't belong to topic you are replying to.")
+				messages.warning(request, "You tried to quote a post which "
+					"doesn't exist or it doesn't belong to topic you are "
+					"replying to. Nice try, Kevin.")
 	return {'topic': topic, 'form': form}
 
 
 @login_required
 @render_to('forums/post_edit.html')
 def post_edit(request, post_id):
-	post = get_object_or_404(Post, pk=post_id)
-	if post.topic.is_closed and not request.user.is_staff:
-		messages.error(request, "You are not allowed to edit posts \
-			in closed topics.")
-		return redirect(topic.get_absolute_url())
-	if not editable_by(post, request.user):
-		messages.error(request, "You are not allowed to edit this post.")
-		return redirect(topic.get_absolute_url())
+	post = get_object_or_404(Post.objects.select_related(), pk=post_id)
+	if not request.user.is_staff:
+		if post.topic.is_closed:
+			messages.error(request, "You are not allowed to edit posts \
+				in closed topics.")
+			return redirect(topic.get_absolute_url())
+		if not post.author == request.user:
+			messages.error(request, "You are not allowed to edit this post.")
+			return redirect(topic.get_absolute_url())
 	if request.method == 'POST':
 		form = PostForm(request.POST)
 		if form.is_valid():
@@ -178,10 +197,13 @@ def post_edit(request, post_id):
 	return {'post': post, 'form': form}
 
 
-@has_perm_or_403('forums.delete_post')
+@login_required
 @render_to('forums/post_delete.html')
 def post_delete(request, post_id):
 	post = get_object_or_404(Post.objects.select_related(), pk=post_id)
+	if not request.user.is_staff:
+		messages.error(request, "You are not allowed to delete this post.")
+		return redirect(topic.get_absolute_url())
 	if request.method == 'POST':
 		if not 'cancel' in request.POST and 'confirmation' in request.POST:
 			post.delete()
@@ -200,26 +222,6 @@ def post_delete(request, post_id):
 			Be ABSOLUTELY sure what you are doing, because this action \
 			cannot be reverted.")
 	return {'post': post, 'form': form}
-
-
-@render_to('forums/search_form.html')
-def search(request):
-	if request.method == 'POST':
-		form = PostSearchForm(request.POST)
-		if form.is_valid():
-			if form.cleaned_data['show_as'] == 'posts':
-				queryset = Post.objects.filter(
-					content__contains=form.cleaned_data['keywords'])
-			else:
-				queryset = Topic.objects.filter(
-					Q(posts__content__contains=form.cleaned_data['keywords']) |
-					Q(title__contains=form.cleaned_data['keywords'])
-				)
-			return {'TEMPLATE': 'forums/search_results.html',
-				'queryset': queryset, 'form': form}
-	else:
-		form = PostSearchForm()
-	return {'form': form}
 
 
 @login_required
@@ -243,48 +245,40 @@ def post_report(request, post_id):
 @user_passes_test_or_403(lambda u: u.is_staff)
 def close_topic(request, topic_id):
 	topic = get_object_or_404(Topic, pk=topic_id)
-	if topic.is_closed:
-		messages.info(request, "Topic is already closed!")
-	else:
-		topic.is_closed = True
-		topic.save()
-		messages.success(request, "Closed.")
+	topic.is_closed = True
+	topic.save()
+	cache.set('forums_topic_%s' % topic_id, topic)
+	messages.success(request, "Closed.")
 	return redirect(topic.get_absolute_url())
 
 
 @user_passes_test_or_403(lambda u: u.is_staff)
 def open_topic(request, topic_id):
 	topic = get_object_or_404(Topic, pk=topic_id)
-	if not topic.is_closed:
-		messages.info(request, "Topic is already open!")
-	else:
-		topic.is_closed = False
-		topic.save()
-		messages.success(request, "Opened.")
+	topic.is_closed = False
+	topic.save()
+	cache.set('forums_topic_%s' % topic_id, topic)
+	messages.success(request, "Opened.")
 	return redirect(topic.get_absolute_url())
 
 
 @user_passes_test_or_403(lambda u: u.is_staff)
 def stick_topic(request, topic_id):
 	topic = get_object_or_404(Topic, pk=topic_id)
-	if topic.is_sticky:
-		messages.info(request, "Topic is already sticked!")
-	else:
-		topic.is_sticky = True
-		topic.save()
-		messages.success(request, "Sticked.")
+	topic.is_sticky = True
+	topic.save()
+	cache.set('forums_topic_%s' % topic_id, topic)
+	messages.success(request, "Sticked.")
 	return redirect(topic.get_absolute_url())
 
 
 @user_passes_test_or_403(lambda u: u.is_staff)
 def unstick_topic(request, topic_id):
 	topic = get_object_or_404(Topic, pk=topic_id)
-	if not topic.is_sticky:
-		messages.info(request, "Topic is already unsticked!")
-	else:
-		topic.is_sticky = False
-		topic.save()
-		messages.success(request, "Unsticked.")
+	topic.is_sticky = False
+	topic.save()
+	cache.set('forums_topic_%s' % topic_id, topic)
+	messages.success(request, "Unsticked.")
 	return redirect(topic.get_absolute_url())
 
 
@@ -307,6 +301,7 @@ def move_topic(request, topic_id):
 		form = MoveTopicForm()
 	return {'topic': topic, 'form': form}
 
+
 def post_vote(request, post_id, value):
 	post = get_object_or_None(Post, pk=post_id)
 	if post == None:
@@ -322,7 +317,7 @@ def post_vote(request, post_id, value):
 			This actually works, so don't try to fix it.
 
 			No really, don't.
-			
+
 			If you still don't believe me, the line below
 			subtracts current karma value and adds new one.
 		"""
@@ -331,6 +326,7 @@ def post_vote(request, post_id, value):
 		karma.save()
 	return post
 
+
 @login_required
 def post_voteup(request, post_id):
 	result = post_vote(request, post_id, 1)
@@ -338,7 +334,9 @@ def post_voteup(request, post_id):
 		messages.error(request, result[0])
 		if result[1]:
 			return redirect(reverse('forums.views.index'))
+	messages.info(request, "Saved.")
 	return post_permalink(request, post_id)
+
 
 @login_required
 def post_votedown(request, post_id):
@@ -347,4 +345,5 @@ def post_votedown(request, post_id):
 		messages.error(request, result[0])
 		if result[1]:
 			return redirect(reverse('forums.views.index'))
+	messages.info(request, "Saved.")
 	return post_permalink(request, post_id)
